# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Delilah_resto",
    "description": "Food Restaurant API",
    "contact": { "email": "jta1223@gmail.com" },
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": ["http"],
  "paths": {
...
    "/user/{id}": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["User"],
        "summary": "List User",
        "operationId": "listUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "type": "integer",
            "name": "id",
            "description": "User Id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/listUser"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["User"],
        "summary": "Updated User",
        "operationId": "udpatedUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "type": "integer",
            "name": "id",
            "description": "User Id",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "description": "Updated user",
            "required": true,
            "schema": {
              "$ref": "#/definitions/updatedUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/listUser"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["User"],
        "summary": "Deleted User",
        "operationId": "deletedUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "type": "integer",
            "name": "id",
            "description": "User Id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/user/{id}/enable": {
      "put": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["User"],
        "summary": "Enable User",
        "operationId": "enableUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "type": "integer",
            "name": "id",
            "description": "User Id",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "description": "Enable user",
            "required": true,
            "schema": {
              "$ref": "#/definitions/enableUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/listUser"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Enter your bearer token in the format JWT"
    }
  },
  "definitions": {
...
    "listUser": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32",
          "example": 1
        },
        "username": {
          "type": "string",
          "example": "Jonathan Jimenez S"
        },
        "fullname": {
          "type": "string",
          "example": "jta1223"
        },
        "email": {
          "type": "string",
          "example": "admin@admin.com"
        },
        "phone": {
          "type": "string",
          "example": "1234567890"
        },
        "address": {
          "type": "string",
          "example": "Cl 10 50 90"
        },
        "admin": {
          "type": "integer",
          "format": "int32",
          "example": 1
        },
        "enable": {
          "format": "int32",
          "type": "integer",
          "example": 1
        },
        "createAt": {
          "type": "string",
          "example": "2020-12-05T02:08:39.000Z"
        }
      }
    },
...
    "enableUser": {
      "type": "object",
      "properties": {
        "enable": {
          "type": "integer",
          "format": "int32",
          "example": 1
        }
      }
    },
...
    "updatedUser": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string",
          "example": "Jonathan Jimenez S"
        },
        "fullname": {
          "type": "string",
          "example": "jta1223"
        },
        "email": {
          "type": "string",
          "example": "admin@admin.com"
        },
        "phone": {
          "type": "string",
          "example": "1234567890"
        },
        "address": {
          "type": "string",
          "example": "Cl 10 50 90"
        }
      }
    },
...
  },
...
}